+++ src/index.html
@@ -9,15 +9,11 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Social Network</title>
   <link rel="stylesheet" href="style.css" />
-  <!--FIRESTORE-->
-  <!-- <script src="https://www.gstatic.com/firebasejs/7.2.3/firebase-app.js"></script>
-  </script> -->
-
+ 
 </head>
 
 <body>
 
-  <img class="logo" src="img/logo.png">
   <div id="root"> </div>
 
   <header id="root2">
@@ -51,8 +47,8 @@
   </script>
 
   <script type="module"src = "main.js" ></script>
-  <script type="module" src="recover.js"></script>
-  <script type="module" src="./lib/post.js"></script>
+ 
+
